requirements.txt
PyGithub==2.1.1
pandas==2.1.4
python-dotenv==1.0.0
notion-client==2.2.1
